Refactor the main CLI command into its own file
